WIP:; Refactoring and Unit Tests
